Firebase integration with viewmodel

+++ app/src/main/java/io/github/notefydadm/notefy/model/CheckBoxBlock.java
@@ -33,7 +33,7 @@ public class CheckBoxBlock extends TextBlock {
         this.isChecked = isChecked;
     }
 
-    public CheckBoxBlock(String text, boolean isChecked, String fontFamily, float fontSize, String textStyle) {
+    public CheckBoxBlock(String text, boolean isChecked, String fontFamily, double fontSize, String textStyle) {
         super(text, fontFamily, fontSize, textStyle);
         this.isChecked = isChecked;
     }
@@ -50,4 +50,5 @@ public class CheckBoxBlock extends TextBlock {
     public String getContent() {
         return "- [" + (isChecked ? "x" : " ") + "] " + super.getContent();
     }
+
 }
