+++ app/src/main/java/io/github/notefydadm/notefy/model/TextBlock.java
@@ -8,17 +8,18 @@ public class TextBlock implements Block {
     private String textStyle;
 
     @Dimension(unit = Dimension.SP)
-    private float fontSize;
+    private double fontSize;
 
     public TextBlock(String text) {
         this(text, "sans-serif", 18, "normal");
     }
 
-    public TextBlock(String text, String fontFamily, float fontSize, String textStyle) {
+    public TextBlock(String text, String fontFamily, double fontSize, String textStyle) {
         this.text = text;
         this.fontFamily = fontFamily;
         this.fontSize = fontSize;
         this.textStyle = textStyle;
+
     }
 
     public String getText() {
@@ -37,7 +38,7 @@ public class TextBlock implements Block {
         this.fontFamily = fontFamily;
     }
 
-    public float getFontSize() {
+    public double getFontSize() {
         return fontSize;
     }
 
@@ -57,4 +58,6 @@ public class TextBlock implements Block {
     public String getContent() {
         return this.getText();
     }
+
+
 }
